pyproject.toml
[project]
name = "golfbackend"
version = "0.1.0"
description = "Golf tee time booking backend"
requires-python = ">=3.10"
dependencies = [
    "django==5.0.6",
    "requests==2.32.3",
    "djangorestframework==3.14.0",
    "beautifulsoup4==4.12.3",
    "django-cors-headers==4.3.1",
    "geopy==2.4.1",
    "haversine==2.3.1",
    "openai==1.69.0",
    "uwsgi==2.0.28",
    "flake8>=7.2.0",
]

[project.optional-dependencies]
dev = [
    "black==24.3.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 120
preview = true

[tool.blackd]
host = "localhost"
port = 45484

[dependency-groups]
dev = [
    "ruff>=0.11.6",
]
